renamed new_tracker to tracker, main ui/config update
