+++ config/oot_keysanity/config.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Root>
-    <Config DefaultInventory="0" StatePath="config/oot/state.txt" ShowTimer="True">
+    <Config XMLVersion="1.0" Name="Ocarina of Time - Keysanity" Icon="items/ocarina_of_time.png" DefaultInventory="0" StatePath="config/oot/state.txt" ShowTimer="False">
         <Fonts>
             <Item Index="0" Name="LEMON MILK Bold" Source="fonts/LEMONMILK-Bold.otf"/>
             <Item Index="1" Name="Upheaval TT (BRK)" Source="fonts/upheavtt.ttf"/>
@@ -18,6 +18,7 @@
             <Item Index="5" Name="LiveSplit" FontIndex="4" Size="40" Bold="True" Color="0xA0A0A0" ColorAlt="0x000000" IsTimer="True" UseGradient="True" Minimal="True"/>
             <Item Index="6" Name="Small Key Counters" FontIndex="2" Size="13.0" Bold="True" Color="0xFFFFFF" ColorAlt="0x00FF00" OutlineThickness="2"/>
             <Item Index="7" Name="Small Key Labels" FontIndex="1" Size="9.0" Bold="True" Color="0xFFFFFF" ColorAlt="0x000000" OutlineThickness="2"/>
+            <Item Index="8" Name="Keysanity" FontIndex="0" Size="11.0" Bold="True" Color="0xFFFFFF" ColorAlt="0x000000" OutlineThickness="2"/>
         </TextSettings>
         <Flags>
             <!-- the flags' positions are relative to the item there are associate to -->
@@ -228,52 +229,53 @@
             <Item ExtraIndex="0" Pos="258;352" Name="Prelude of Light" Source="items/songs/prelude.png"/>
 
             <!-- Keys -->
+            <Text Index="0" Pos="288;408" Width="72" Height="12" Rot="90" Content="Dungeons" TextSettings="8"/>
             <Item Pos="8;400" Name="Forest Temple Small Key" Source="items/small_key.png" ScaleContent="True" UseWheel="True">
-                <Text Index="0" Pos="25;395" Width="48" Height="12" Content="Forest" TextSettings="7"/>
+                <Text Index="1" Pos="25;395" Width="48" Height="12" Rot="0" Content="Forest" TextSettings="7"/>
                 <Counter TextSettings="6" Min="1" Max="5" Increment="1" Pos="10;22" Width="32" Height="15"/>
             </Item>
             <Item Pos="49;400" Name="Forest Temple Boss Key" Source="items/boss_key.png" ScaleContent="True">
             </Item>
             <Item Pos="90;400" Name="Shadow Temple Small Key" Source="items/small_key.png" ScaleContent="True" UseWheel="True">
-                <Text Index="0" Pos="100;395" Width="54" Height="12" Content="Shadow" TextSettings="7"/>
+                <Text Index="2" Pos="100;395" Width="54" Height="12" Rot="0" Content="Shadow" TextSettings="7"/>
                 <Counter TextSettings="6" Min="1" Max="5" Increment="1" Pos="10;22" Width="32" Height="15"/>
             </Item>
             <Item Pos="131;400" Name="Shadow Temple Boss Key" Source="items/boss_key.png" ScaleContent="True">
             </Item>
             <Item Pos="8;441" Name="Fire Temple Small Key" Source="items/small_key.png" ScaleContent="True" UseWheel="True">
-                <Text Index="0" Pos="25;435" Width="36" Height="12" Content="Fire" TextSettings="7"/>
+                <Text Index="3" Pos="25;435" Width="36" Height="12" Rot="0" Content="Fire" TextSettings="7"/>
                 <Counter TextSettings="6" Min="1" Max="8" Increment="1" Pos="10;22" Width="32" Height="15"/>
             </Item>
             <Item Pos="49;440" Name="Fire Temple Boss Key" Source="items/boss_key.png" ScaleContent="True">
             </Item>
             <Item Pos="90;441" Name="Spirit Temple Small Key" Source="items/small_key.png" ScaleContent="True" UseWheel="True">
-                <Text Index="0" Pos="100;435" Width="48" Height="12" Content="Spirit" TextSettings="7"/>
+                <Text Index="4" Pos="100;435" Width="48" Height="12" Rot="0" Content="Spirit" TextSettings="7"/>
                 <Counter TextSettings="6" Min="1" Max="5" Increment="1" Pos="10;22" Width="32" Height="15"/>
             </Item>
             <Item Pos="131;440" Name="Spirit Temple Boss Key" Source="items/boss_key.png" ScaleContent="True">
             </Item>
             <Item Pos="8;481" Name="Water Temple Small Key" Source="items/small_key.png" ScaleContent="True" UseWheel="True">
-                <Text Index="0" Pos="25;475" Width="45" Height="12" Content="Water" TextSettings="7"/>
+                <Text Index="5" Pos="25;475" Width="45" Height="12" Rot="0" Content="Water" TextSettings="7"/>
                 <Counter TextSettings="6" Min="1" Max="6" Increment="1" Pos="10;22" Width="32" Height="15"/>
             </Item>
             <Item Pos="49;481" Name="Water Temple Boss Key" Source="items/boss_key.png" ScaleContent="True">
             </Item>
             <Item Pos="90;481" Name="Ganon Small Key" Source="items/small_key.png" ScaleContent="True" UseWheel="True">
-                <Text Index="0" Pos="100;475" Width="45" Height="12" Content="Ganon" TextSettings="7"/>
+                <Text Index="6" Pos="100;475" Width="45" Height="12" Rot="0" Content="Ganon" TextSettings="7"/>
                 <Counter TextSettings="6" Min="1" Max="1" Increment="1" Pos="10;22" Width="32" Height="15"/>
             </Item>
             <Item Pos="132;481" Name="Ganon Boss Key" Source="items/boss_key.png" ScaleContent="True">
             </Item>
             <Item Pos="194;400" Name="BOTW Small Key" Source="items/small_key.png" ScaleContent="True" UseWheel="True">
-                <Text Index="0" Pos="188;395" Width="45" Height="12" Content="BOTW" TextSettings="7"/>
+                <Text Index="7" Pos="188;395" Width="45" Height="12" Rot="0" Content="BOTW" TextSettings="7"/>
                 <Counter TextSettings="6" Min="1" Max="3" Increment="1" Pos="10;22" Width="32" Height="15"/>
             </Item>
             <Item Pos="194;441" Name="GTG Small Key" Source="items/small_key.png" ScaleContent="True" UseWheel="True">
-                <Text Index="0" Pos="188;435" Width="45" Height="12" Content="GTG" TextSettings="7"/>
+                <Text Index="8" Pos="188;435" Width="45" Height="12" Rot="0" Content="GTG" TextSettings="7"/>
                 <Counter TextSettings="6" Min="1" Max="9" Increment="1" Pos="10;22" Width="32" Height="15"/>
             </Item>
             <Item Pos="194;481" Name="Fortress Small Key" Source="items/small_key.png" ScaleContent="True" UseWheel="True">
-                <Text Index="0" Pos="175;475" Width="72" Height="12" Content="Fortress" TextSettings="7"/>
+                <Text Index="9" Pos="175;475" Width="72" Height="12" Rot="0" Content="Fortress" TextSettings="7"/>
                 <Counter TextSettings="6" Min="1" Max="4" Increment="1" Pos="10;22" Width="32" Height="15"/>
             </Item>
         </Inventory>
